requirements.txt
langchain
langchain-openai
dotenv
openai
groq
google-generative-ai
